% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/even-odd-no-check.R
\name{check_even_odd}
\alias{check_even_odd}
\title{Check Even or Odd}
\usage{
check_even_odd(n)
}
\arguments{
\item{n}{An integer.}
}
\value{
A character indicating "even" or "odd".
}
\description{
This function checks whether a given number is even or odd.
}
\examples{
check_even_odd(5)


}
