% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fibonacci_seq-calculator.R
\name{fibonacci_sequence}
\alias{fibonacci_sequence}
\title{Fibonacci Sequence}
\usage{
fibonacci_sequence(n)
}
\arguments{
\item{n}{An integer specifying the number of terms in the sequence.}
}
\value{
A numeric vector containing the Fibonacci sequence.
}
\description{
This function generates the Fibonacci sequence up to a specified number of terms.
}
\examples{
fibonacci_sequence(10)


}
