% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rm_0s_by_column.R
\name{rm_0s_by_col}
\alias{rm_0s_by_col}
\title{Removing zeros from a specified matrix by column}
\usage{
rm_0s_by_col(mat)
}
\arguments{
\item{mat}{Input matrix}
}
\value{
Matrix with rows containing all 0s removed
}
\description{
This function removes all zeros by column from the matrix;
`set.seed(123)`
`x = matrix(rpois(100, 0.1), 10, 10)`
}
\examples{
mat <- matrix(c(0, 1, 0, 0, 0, 2, 0, 3, 0), ncol = 3)
rm_0s_by_col(mat)
}
